Refactoring file structure again, some renaming, some fixing etc.
