+++ bot/handlers/random_content.py
